pyproject.toml
[tool.poetry]
name = "jetscape_analysis"
version = "0.1.0"
description = "Basic JETSCAPE analysis code"
license = "BSD-3-Clause"
authors = [
    "James Mulligan <james.mulligan@berkeley.edu>",
    "Raymond Ehlers <raymond.ehlers@cern.ch>",
]
classifiers = [
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    'Development Status :: 4 - Beta',
    # Indicate who your project is intended for
    'Intended Audience :: Science/Research',
    # Specify the Python versions you support here. In particular, ensure
    # that you indicate whether you support Python 2, Python 3 or both.
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
tqdm = "^4.37.0"
numpy = "^1.17.3"
pyhepmc_ng = "^0.4.3"
pyyaml = "^5.1.2"
isort = {extras = ["pyproject"], version = "^4.3.21"}
awkward = "^1.2"
pyarrow = "^3.0.0"
pandas = "^1.2.4"
attrs = "^20.3.0"
numba = "^0.53.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2.2"
flake8 = "^3.7.9"
flake8-colors = "^0.1.6"
black = "^19.10b0"
pre-commit = "^1.20.0"
mypy = "^0.812"

[tool.isort]
known_standard_library = ["setuptools"]
known_third_party = ["ROOT", "common_base", "event", "fastjet", "fjext", "pyhepmc_ng", "reader", "tqdm", "yaml"]
known_first_party = ["heppy"]
line_length = 120
# Meaning a grouped grid of imports
multi_line_output = 3
lines_after_imports = 1
force_grid_wrap = 0
use_parentheses = true
verbose = true
skip = ["docs/conf.py"]
skip_glob = ".*"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

